requirements.txt
Django==4.2.17
Pillow==10.2.0
django-cors-headers==4.3.1
django-htmlmin==0.11.0
django-meta==2.4.1
django-robots==6.1
django-seo-js==0.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
asgiref==3.8.1
sqlparse==0.5.3
google-api-python-client==2.100.0
requests==2.31.0
